### 2019-06-20 #1514 @danielskatz
@xoolive - please review and update the bib entries - as you can see in the pdf, there are some issues.  You can wrap words or letters in {}s to preserve their case, for example. And the third reference is incomplete.  Once you have made changes in the .bib file, you can regenerate the pdf by adding a new comment in this issue with ```@whedon generate pdf```
