+++ joss/paper.bib
@@ -21,23 +21,22 @@
 @inproceedings{olive_detecting_2018,
 	title = {Detecting {Controllers}’ {Actions} in {Past} {Mode} {S} {Data} by {Autoencoder}-{Based} {Anomaly} {Detection}},
 	author = {Olive, Xavier and Grignard, Jeremy and Dubot, Thomas and Saint-Lot, Julie},
-	booktitle = {Proceedings of the SESAR Innovation Days},
+	booktitle = {Proceedings of the {SESAR} {Innovation} {Days}},
 	year = {2018},
 	pages = {8},
 }
 
 @inproceedings{olive_identifying_2019,
 	title = {Identifying {Anomalies} in past en-route {Trajectories} with {Clustering} and {Anomaly} {Detection} {Methods}},
-	booktitle = {Proceedings of the {USA/FAA} {Air} {Traffic} {Management} {R&D} {Seminar} 2019},
+	booktitle = {Proceedings of the {Air} {Traffic} {Management} {R&D} {Seminar}},
 	author = {Olive, Xavier and Basora, Luis},
 	year = {2019},
-	pages = {10},
 }
 
 @inproceedings{schafer_opensky_2019,
 	title = {{OpenSky} {Report} 2019: {Analysing} {TCAS} in the {Real} {World} using {Big} {Data}},
-	booktitle = {Digital {Avionics} {Systems} {Conference} ({DASC}), 2019 {IEEE}/{AIAA} 38th},
-	author = {Schafer, Matthias and Olive, Xavier and Strohmeier, Martin and Smith, Matthew and Martinovic, Ivan and Lenders, Vincent},
+	booktitle = {Proceedings of the 38th {Digital} {Avionics} {Systems} {Conference} ({DASC})},
+	author = {Schäfer, Matthias and Olive, Xavier and Strohmeier, Martin and Smith, Matthew and Martinovic, Ivan and Lenders, Vincent},
 	year = {2019},
 }
 
@@ -51,7 +50,7 @@
 @article{sun_pymodes_2019,
     author={J. {Sun} and H. {V\^u} and J. {Ellerbroek} and J. M. {Hoekstra}},
     journal={IEEE Transactions on Intelligent Transportation Systems},
-    title={pyModeS: Decoding Mode-S Surveillance Data for Open Air Transportation Research},
+    title={pyModeS: Decoding {Mode S} {Surveillance} {Data} for {Open} {Air} {Transportation} {Research}},
     year={2019},
     doi={10.1109/TITS.2019.2914770},
     ISSN={1524-9050},
@@ -61,7 +60,7 @@
   author =    {Sean Gillies and others},
   organization = {toblerity.org},
   title =     {Shapely: manipulation and analysis of geometric objects},
-  year =      {2007--},
+  year =      {2007},
   url = "https://github.com/Toblerity/Shapely"
 }
 
@@ -72,13 +71,13 @@
   month = {dec},
   publisher = {The Open Journal},
   author = {Jacob VanderPlas and Brian Granger and Jeffrey Heer and Dominik Moritz and Kanit Wongsuphasawat and Arvind Satyanarayan and Eitan Lees and Ilia Timofeev and Ben Welsh and Scott Sievert},
-  title = {Altair: Interactive Statistical Visualizations for Python},
+  title = {Altair: Interactive {Statistical} {Visualizations} for {Python}},
   journal = {Journal of Open Source Software}
 }
 
 @misc{cartopy_2018,
[…]
