+++ joss/paper.bib
@@ -28,7 +28,7 @@
 
 @inproceedings{olive_identifying_2019,
 	title = {Identifying {Anomalies} in past en-route {Trajectories} with {Clustering} and {Anomaly} {Detection} {Methods}},
-	booktitle = {Proceedings of the {Air} {Traffic} {Management} {R&D} {Seminar}},
+	booktitle = {Proceedings of the {Air} {Traffic} {Management} {Seminar}},
 	author = {Olive, Xavier and Basora, Luis},
 	year = {2019},
 }
