### 2019-06-21 #1514 @espinielli
>
>
> 👋 @junzis & @espinielli - would one of you be willing to review this submission for [JOSS](http://joss.theoj.org/about#about)? A JOSS review involves examining the short paper and the software repository against [18 criteria](https://joss.readthedocs.io/en/latest/review_checklist.html), and explaining problems to the editor and to the author, who makes improvements in an iterative process until the submission is acceptable.

I would be happy. What's the deadline? (leaving sooooon for a week...)
