### 2019-06-21 #1518 @whedon
[REVIEW]: traffic: a toolbox for processing and analysing air traffic data
**Submitting author:** @xoolive (<a href="http://orcid.org/0000-0002-2335-5774">Xavier Olive</a>)
**Repository:** <a href="https://github.com/xoolive/traffic" target ="_blank">https://github.com/xoolive/traffic</a>
**Version:** v2.0
