### 2019-06-21 #1518 @danielskatz
Thanks for agreeing to review @junzis & @espinielli - instructions are above, which are basically to work through your checklists.  I'lll also note here that @espinelli will be out for about a week soon.

If you have any questions, please ask.
