### 2019-06-23 #1518 @espinielli
I, too, am writing (and presenting at [UseR!2019](http://www.user2019.fr): "R in the Air") an R package for flight trajectories 😇
