### 2019-06-25 #1518 @junzis
@danielskatz - Do items from the checklist "Document" section also need to be included in the paper? Thanks!
