### 2019-06-25 #1518 @danielskatz
The documentation criteria is:
> There should be sufficient documentation for you, the reviewer to understand the core functionality of the software under review. A high-level overview of this documentation should be included in a README file (or equivalent).

The paper needs are shown at the top of https://joss.readthedocs.io/en/latest/review_criteria.html
