### 2019-06-25 #1518 @junzis
@xoolive - The paper, library, and document all look great! Here are my minor comments:

1) It will be great if you can emphasize "the statement of need" in the paper. Perhaps from the perspective of ATM researcher point of view.

2) Is there already an archive version of the library on Zenodo, Figshare, or similar?  The "Archive" link in the PDF seems not working.
