### 2019-06-25 #1518 @xoolive
Thank you @junzis for the feedback.

1. I will emphasize this statement of need and let you know here when ready.
2. It looks like it's all [there](https://joss.readthedocs.io/en/latest/whedon.html?highlight=zenodo#setting-the-software-archive)!
3. I am not sure it is a good idea: pyarrow and fastparquet are optional dependencies for pandas for a reason. Both have strengths and weaknesses and the decision can/should be let to the end user (Parquet format is not a must-go for the typical user of this library). From the top of my head, I think the HDF5 engine also requires optional dependencies which are useless for users like myself who almost never use the HDF5 format.
  Error messages of pandas for optional dependencies are rather explicit. It seems to me that questioning their dependency choices is out of the scope of this library.
