### 2019-06-27 4ac5d96
joss review: comment on statement of need by @junzis
