### 2019-06-27 #1518 @xoolive
@junzis Latest commit addresses your comment about the statement of need.
https://github.com/xoolive/traffic/commit/4ac5d962f4b8978cc704a7f57ca9b53910654cc9
