### 2019-07-04 #1518 @espinielli
@xoolive I am trying to install as per https://traffic-viz.github.io/installation.html
Specifically:
1. clone of the repo
1. `python setup.py install`

Last failed: I am on a Mac where python is still 2.7.0, ok, so I installed "latest" `python` (`brew install python`)
and run `python3 setup.py install` which fails with

```
...
    exec(code, globals, locals)
  File "/var/folders/bw/cfz6xpw10cs4l_1jkhyyywpw0000gn/T/easy_install-hl3fy6pw/Cartopy-0.17.0/setup.py", line 42, in <module>
    file.relative_to(sample_dir).as_posix()
ImportError: Cython 0.15.1+ is required to install cartopy.
```

(I assume Python 3 is the norm these days, hence why you did not mentioned it in the documentstion)
