### 2019-07-04 xoolive/traffic#28 @espinielli
JOSS review: rewording
I think in joss/paper.md:

```
... indexed by time, which derivate as ground speed, track ...
```
should be reworded to something like

```
... indexed by time, with derived ones such as ground speed, track ...
```
