### 2019-07-04 xoolive/traffic#28 @espinielli
Also

```
...
Further, airspaces are a key element of aviation: they are regulated by specific rules, navigation is allowed to determined types of aircraft meeting strict requirements. Such areas ...
```

could be changed to something like:

```
...
Further, airspaces are a key element of aviation: they are regulated by specific rules, whereby navigation is allowed for determined types of aircraft meeting strict requirements. Such volumes ...
```
