### 2019-07-04 xoolive/traffic#29 @espinielli
JOSS review: reference to Python, Pandas, Shapely seems isolated
I think the message being passed about Python, Pandas and Shapely is that these are used by `traffic`, so the paragraph

```
Python is a popular language among academics and data scientists for its active
 community and large catalogue of libraries fitting the needs of most in their
respective areas of research.
Pandas is a natural solution to represent time series,
 Shapely [@shapely_2007] a package of choice to address geometrical operations
(intersections, inclusion, cascaded joins among others).
```
is coming IMHO too early...
Maybe something like the following will make that para less standalone:

```
The traffic library uses the Python language to reach to its large base of academics and data scientists, to serve and benefit its active community and to build on top of a
 large catalogue of libraries.
So trajectories are modeled on top of Pandas dataframe, a natural solution to
 represent time series,  while airspaces delegate to Shapely [@shapely_2007] geometrical operations (intersections, inclusion, cascaded joins among others).
```
