### 2019-07-04 #1518 @espinielli
Small suggestion for rewording in https://github.com/xoolive/traffic/issues/28
Small suggestion for reordering of text https://github.com/xoolive/traffic/issues/29
