### 2019-07-04 #1518 @xoolive
Nothing stupid @espinielli, just regular Python mess 😄

Python 3 is indeed the norm these days, and the dependency is stated [here](https://github.com/xoolive/traffic/blob/master/setup.py#L51). The installation process should fail if you have python<3.6.

There are specific non-Python dependencies for two Python libraries, namely `cartopy` and `shapely`. For Mac OS, brew is an option but you need to pay extra attention to the cross dependencies between the versions of the C libraries and of the Python libraries. Then `Cython` is indeed required to build `cartopy` but not to execute it.

If you already have a working `cartopy` and `shapely`, then installation is direct. Otherwise, as recommended [here](https://traffic-viz.github.io/installation.html), it may be better to use an anaconda distribution which installs the C libraries and the Python libraries, and also does the dependency mess for you!
