### 2019-07-04 025b0b8
joss review: rewording by @espinielli fix #28
