### 2019-07-04 d953588
joss review: rephrasing/reordering by @espinielli fix #29
