### 2019-07-04 xoolive/traffic#29 @xoolive
Thank you for the suggestion @espinielli
I reworked the two paragraphs to make it sound
