### 2019-07-05 xoolive/traffic#30 @espinielli
JOSS review: sample data in quick start
When following the Quick Start steps I found that if the claim is to start from the cloned repo
I have to correct the path to the sample data from

```
from traffic.core import Traffic
t = Traffic.from_file('../data/sample_opensky.pkl')
```

to (one dot, i.e. `data` directory under current working directory)

```
from traffic.core import Traffic
t = Traffic.from_file('./data/sample_opensky.pkl')
```
