### 2019-07-05 xoolive/traffic#31 @espinielli
JOSS Review: access to NM AIRAC data not clear
The part of the QS that says

If you have access to AIRAC files (published by Eurocontrol), you may get information about more complex zones, here the Terminal Maneuvering Area (TMA) for Toulouse.
```python
from traffic.data import nm_airspaces
nm_airspaces['LFBOTMA']
```
fails for me...(and subsequent ones where `nm_airspaces` is used)
Does it need some additional data sources?
Probably a link to [Data Sources](https://traffic-viz.github.io/airac_usage.html) would help...
