### 2019-07-05 xoolive/traffic#31 @xoolive
Thank your for the extensive feedback and for pointing at such dark corners of the documentation!

How does it fail for you? Do you have a proper error message suggesting to edit the configuration file? Do you mean you would like to see the link to the documentation page in the exception or do you mean it would be better to add this link in the quickstart page? or both?
