### 2019-07-05 xoolive/traffic#30 @xoolive
Thank you for pointing at this. Correcting the update path on other occurrences as well.
