### 2019-07-05 xoolive/traffic#31 @espinielli
I think it would be good to mention the link in the Quick Start documentation.
