### 2019-07-05 xoolive/traffic#31 @espinielli
If I run

```python
from traffic.data import nm_airspaces
nm_airspaces['LFBOTMA']
```

I get

> ---------------------------------------------------------------------------
> RuntimeError                              Traceback (most recent call last)
> <ipython-input-11-76c3defdfd9c> in <module>
>       1 from traffic.data import nm_airspaces
> ----> 2 nm_airspaces['LFBOTMA']
>
> ~/kaos/traffic/traffic/data/airspaces/eurocontrol_nm.py in __getitem__(self, name)
>     120
>     121         if not self.initialized:
> --> 122             self.init_cache()
>     123
>     124         list_names = self.airspaces.get(name, None)
>
> ~/kaos/traffic/traffic/data/airspaces/eurocontrol_nm.py in init_cache(self)
>      35
>      36         if self.nm_path is None:
> ---> 37             raise RuntimeError(msg)
>      38
>      39         are_file = next(self.nm_path.glob("Sectors_*.are"), None)
>
> RuntimeError: Edit file /Users/espin/Library/Application Support/traffic/traffic.conf with NM directory
>
>

which clearly says to properly edit your config directory, but at this stage of the QS you do not even have any idea about the airspace data, hence a link to the relevant section in the docs would be beneficial...
