### 2019-07-05 #1518 @espinielli
I have finished my review. The package and the paper are excellent.
I have filed 2 minor suggestion on the documentation which @xoolive is addressing faster than
light speed 👍 https://github.com/xoolive/traffic/issues/30 and https://github.com/xoolive/traffic/issues/31.

IMHO they are not to be blocking issues for the paper to be accepted.
