### 2019-07-05 ad1736d
joss review: link to NM configuration @espinielli fix #31
