### 2019-07-05 xoolive/traffic#31 @xoolive
Added a warning block at this stage of the documentation with the relevant link. Thanks for pointing at it!
