### 2019-07-05 #1518 @danielskatz
@xoolive - we now need you to archive the software in an archival repository, such as zenodo (see https://guides.github.com/activities/citable-code/) or elsewhere that provides a DOI - then let me know the DOI of the archive.  Also, is 2.0 still the correct version number?  If not, let me know what is.
