### 2019-07-05 #1518 @xoolive
Thank you @danielskatz

Here is the link to the DOI: https://doi.org/10.5281/zenodo.3269812
2.0 is the correct version number
