+++ joss/paper.bib
@@ -42,8 +42,9 @@
 
 @inproceedings{schafer_opensky_2014,
 	title = {Bringing up {OpenSky}: {A} large-scale {ADS}-{B} sensor network for research},
-	booktitle = {Proceedings of the 13th international symposium on {Information} processing in sensor networks},
+	booktitle = {Proceedings of the 13th International Symposium on Information Processing in Sensor Networks},
 	author = {Schäfer, Matthias and Strohmeier, Martin and Lenders, Vincent and Martinovic, Ivan and Wilhelm, Matthias},
+	doi = {10.1109/IPSN.2014.6846743},
 	year = {2014},
 }
 
