### 2019-07-05 #1518 @xoolive
@danielskatz done, I am not sure now about the whedon command to type in order to generate a new final proof
