### 2019-07-05 #1518 @danielskatz
sorry, my fix didn't quite work for the cases - let's try again - please change the line in the bib file to:
booktitle = {Proceedings of the 13th {I}nternational {S}ymposium on {I}nformation {P}rocessing in {S}ensor {N}etworks},
